Mismatch between C++ and Python for callback index type
When multiple callbacks are defined, callbacks are stored in
an array on both the client and server. When a callback is
registered, only the index of the callback in this array is
sent over the transport. In C++, the array index is encoded as
a uint8, where as in Python it is encoded as a uint32. So, a
Python client will not be aligned with a C++ server, leading
to wrong information being decoded by the server. This fix
aligns both to be a uint8 type for the callback index.