Fix memset past end of buffer in erpc_setup_mbf_static.cpp
The memset of the m_buffers variable was not being done correctly.

For example, consider:
  ERPC_DEFAULT_BUFFERS_COUNT=3
  ERPC_DEFAULT_BUFFER_SIZE=24

ERPC_BUFFER_SIZE_UINT8 would be (24 + 8 - 1) = 31
ERPC_BUFFER_SIZE_UINT64 would be (31 / 8) = 3
The size of the memset would be (3 * 31) = 93
The size of the array would be (8 * 3 * 3) = 72

So in this example, the memset would go (93 - 72) = 21 bytes past the
end of the array.